feat(Pytorch): Move functionality into builders.py
