+++ src/amltk/pytorch/pytorch_example.py
@@ -16,11 +16,12 @@ from torch import nn, optim
 from torch.optim.lr_scheduler import StepLR
 from torchvision import datasets, transforms
 
-from amltk import Component, Metric, Sequential, Choice, Fixed
+from amltk import Component, Metric, Sequential, Choice
 
 # Change this to optuna if you prefer
 # from amltk.optimization.optimizers.optuna import OptunaParser
 from amltk.optimization.optimizers.smac import SMACOptimizer
+from amltk.pytorch.builders import build_model_from_pipeline, MatchDimensions
 
 if TYPE_CHECKING:
     from amltk import Node, Trial
@@ -28,37 +29,10 @@ if TYPE_CHECKING:
 from rich import print
 
 
-class Net(nn.Module):
-    """Reference neural network model."""
-
-    def __init__(self):
-        super().__init__()
-        self.conv1 = nn.Conv2d(1, 32, 3, 1)
-        self.conv2 = nn.Conv2d(32, 64, 3, 1)
-        self.dropout1 = nn.Dropout(0.25)
-        self.dropout2 = nn.Dropout(0.5)
-        self.fc1 = nn.Linear(9216, 128)
-        self.fc2 = nn.Linear(128, 10)
-
-    def forward(self, x):
-        x = self.conv1(x)
-        x = F.relu(x)
-        x = self.conv2(x)
-        x = F.relu(x)
-        x = F.max_pool2d(x, 2)
-        x = self.dropout1(x)
-        x = torch.flatten(x, 1)
-        x = self.fc1(x)
-        x = F.relu(x)
-        x = self.dropout2(x)
-        x = self.fc2(x)
-        return F.log_softmax(x, dim=1)
-
-
 def test(
-    model: nn.Module,
-    device: torch.device,
-    test_loader: torch.utils.data.DataLoader,
+        model: nn.Module,
+        device: torch.device,
+        test_loader: torch.utils.data.DataLoader,
 ) -> tuple[float, float]:
     model.eval()
     test_loss = 0
@@ -76,62 +50,15 @@ def test(
     return float(test_loss), float(accuracy)
 
 
-def build_model_from_pipeline(pipeline: Sequential) -> nn.Module:
-    """Builds a model from the provided pipeline.
-
-    This function iterates through the pipeline nodes, constructing the model
-    layers dynamically based on the node types and configurations. It also
-    utilizes the `MatchDimensions` objects to handle dimension matching
-    between layers with search spaces.
-
-    Args:
-        pipeline: The pipeline containing the model architecture.
-
-    Returns:
-        The constructed PyTorch model.
-    """
-    print("Building model from pipeline")
-    print(pipeline)
-
-    model_layers = []
-
-    for node in pipeline.iter():
-
-        # Check if node is a Flatten layer, ReLU or similar.
-        if isinstance(node, Fixed):
-            model_layers.append( node.item)
-
-        # Check if node is a Component with config parameter
-        elif isinstance(node.item, type) and issubclass(node.item, nn.Module) and node.config:
-            layer_config = node.config or {}
-
-            # Evaluate MatchDimensions objects
-            for key, value in layer_config.items():
-                if isinstance(value, MatchDimensions):
-                    layer_config[key] = value.evaluate(pipeline)
-
-            layer = node.build_item(**layer_config)
-            model_layers.append(layer)
-
-    model = nn.Sequential(*model_layers)
-
-    print("-" * 80)
-    print("Model built")
-    print(model)
-    print("-" * 80)
-
-    return model
-
-
 def eval_configuration(
-    trial: Trial,
-    pipeline: Node,
-    device: str = "cpu",  # Change if you have a GPU
-    epochs: int = 1,  # Fixed for now
-    lr: float = 0.1,  # Fixed for now
-    gamma: float = 0.7,  # Fixed for now
-    batch_size: int = 64,  # Fixed for now
-    log_interval: int = 10,  # Fixed for now
+        trial: Trial,
+        pipeline: Node,
+        device: str = "cpu",  # Change if you have a GPU
+        epochs: int = 1,  # Fixed for now
+        lr: float = 0.1,  # Fixed for now
+        gamma: float = 0.7,  # Fixed for now
+        batch_size: int = 64,  # Fixed for now
+        log_interval: int = 10,  # Fixed for now
 ) -> Trial.Report:
     """Evaluates a configuration within the given trial.
 
@@ -225,42 +152,6 @@ def eval_configuration(
     return trial.success(accuracy=final_test_acc)
 
 
-class MatchDimensions:
-    """Class to handle matching dimensions between layers in a pipeline.
-
-    This class helps ensure compatibility between layers with search spaces
-    during HPO optimization. It takes the layer name and parameter name
-    and stores them for later reference.
-
-    When called, it retrieves the corresponding configuration value from
-    another layer based on the provided information.
-    """
-
-    def __init__(self, layer_name: str, param: str):
-        self.layer_name = layer_name
-        self.param = param
-
-    def evaluate(self, pipeline: Sequential) -> int:
-        """Retrieves the corresponding configuration value from the pipeline.
-
-        Args:
-            pipeline: The pipeline to search for the matching configuration.
-
-        Returns:
-            The value of the matching configuration parameter.
-        """
-        layer = pipeline[self.layer_name]
-        layer_config = layer.config
-        if layer_config is None:
-            raise ValueError(f"Configuration not found for layer {self.layer_name}")
-        value = layer_config.get(self.param)
-        if value is None:
-            raise ValueError(
-                f"Parameter {self.param} not found in config of layer {self.layer_name}",
-            )
-        return value
-
-
 def main() -> None:
     """Main function to orchestrate the model training and evaluation process.
 
@@ -281,21 +172,57 @@ def main() -> None:
 
     # Define the pipeline with search space for hyperparameter optimization
     pipeline = Sequential(
-        nn.Flatten(start_dim=1),
+        Choice(
+
+            Sequential(
+                nn.Flatten(start_dim=1),
+                Component(
+                    nn.Linear,
+                    config={"in_features": 784, "out_features": 100},
+                    # space={"out_features": (100, 500)},
+                    name="choice1-fc1",
+                ),
+                name="choice1",
+            ),
+
+            Sequential(
+                Component(
+                    nn.Conv2d,
+                    config={
+                        "in_channels": 1,  # MNIST images are grayscale, so input channels is 1
+                        "out_channels": 32,  # Number of output channels (filters)
+                        "kernel_size": (3, 3),  # Size of the convolutional kernel
+                        "stride": (1, 1),  # Stride of the convolution
+                        "padding": (1, 1),  # Padding to add to the input
+                    },
+                    name="choice2-conv1",
+                ),
+                nn.ReLU(),
+                nn.MaxPool2d(kernel_size=(2, 2)),
+                nn.Flatten(start_dim=1),
+                name="choice2",
+            ),
+
+            name="layer1",
+        ),
+
         Component(
             nn.Linear,
             config={
-                "in_features": 784,
+                "in_features": 32 * 14 * 14,  # Calculate the input size based on previous layers if Conv2d is chosen
                 "out_features": MatchDimensions("fc2", param="in_features"),
             },
             name="fc1",
         ),
+
         nn.ReLU(),
+
         Component(
             nn.Linear,
             space={"in_features": (10, 50), "out_features": (10, 30)},
             name="fc2",
         ),
+
         Component(
             nn.Linear,
             config={
@@ -304,7 +231,9 @@ def main() -> None:
             },
             name="fc3",
         ),
+
         Component(nn.LogSoftmax, config={"dim": 1}),
+
         name="my-mlp-pipeline",
     )
